+++ react/src/js/components/Consonant/Container/Container.jsx
@@ -909,6 +909,13 @@ const Container = (props) => {
                     }
 
                     setCards(processedCards);
+
+                    // check if the current page is greater than the last page
+                    const lastPage = Math.ceil(processedCards.length / resultsPerPage);
+                    if (currentPage > lastPage) {
+                        setCurrentPage(1);
+                    }
+
                     if (!showEmptyFilters) {
                         setFilters(prevFilters => removeEmptyFilters(prevFilters, processedCards));
                     }
@@ -1453,6 +1460,7 @@ const Container = (props) => {
                             <FiltersPanelTop
                                 filterPanelEnabled={filterPanelEnabled}
                                 filters={filters}
+                                searchQuery={searchQuery}
                                 windowWidth={windowWidth}
                                 resQty={gridCardLen}
                                 onCheckboxClick={handleCheckBoxChange}